+++ packages/style/src/index.ts
@@ -3,6 +3,7 @@
  * @license     https://opensource.org/licenses/MIT
  */
 
+import Renderer from './Renderer';
 import ClientRenderer from './client/ClientRenderer';
 import hydrateStyles from './client/hydrateStyles';
 import ServerRenderer from './server/ServerRenderer';
@@ -12,4 +13,11 @@ import renderToStyleMarkup from './server/renderToStyleMarkup';
 export * from './helpers';
 export * from './types';
 
-export { ClientRenderer, ServerRenderer, captureStyles, hydrateStyles, renderToStyleMarkup };
+export {
+  Renderer,
+  ClientRenderer,
+  ServerRenderer,
+  captureStyles,
+  hydrateStyles,
+  renderToStyleMarkup,
+};